+++ docs/changelog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [Issue #75](https://github.com/gullerya/data-tier/issues/75) - implemented tying declaration API - events
+- documentation updated
+- dependencies maintanance
 
 ## [3.4.0] - 2021-06-15
 ### Changed